Finish up the boot loader proper.
Add the reading in of the secondary boot loader plus analysis of the HOM block,
making sure things are good, decent error reporting and a dead-man timeout in
case no interrupts ever fire (though maybe I can do some of this in a better way
that's shorter w/o establishing the timer ISR, relying on HLT to wake up when a
vertical interrupt happens.

We're right at 511 bytes today. This code is tested through validation of the
HOM block.