New approach to timeouts
Rather than have a countdown with our own IRS, let's cheat a little. Timeouts
don't need to be *THAT* precise, so let's rely on the fact that there's a
veritcal interrupt still, but don't run any code. We also rely on the fact that
hlt will stop the CPU until there's an interrupt. So if we load CX with how long
we want to wait in Hz (always 60 atm), and do a "1: hlt; loop 1b" then we'll
get the timeout w/o the overhead of the ISR, establishing it, tearing it down,
the ISR itself, and the code to look at the variable we're decrementing. Saves
71 bytes w/o losing functionality!

Still not booting secondary program yet for reasons unknown... that's the
next bug to fix... The first sector read is all zeros for some reason (well
the buffer we're supposed to store it in is all 0's!).