Document based on xebec design
The Xebec reference design was used without modification by the Univation disk
controller. When those mnemonics are used for the ports and their bits,
the driver makes a lot of sense.

The driver also assumes that it's talking to an adaptec ACB-4000A which
defines SCSI CMD 0xF to be 'translate' which translates a LBA to CHS
and the driver uses that for error reporting. This isn't a standard
SCSI CMD (0F is used for tapes, but not disks). Document this too.